requirements.txt
#elasticsearch_dsl
numpy
pandas
python-dateutil
opensearch-py
